activity/calculatedelta/activity_test.go
package calculatedelta

import (
	"io/ioutil"
	"testing"

	"github.com/TIBCOSoftware/flogo-contrib/action/flow/test"
	"github.com/TIBCOSoftware/flogo-lib/core/activity"
)

var activityMetadata *activity.Metadata

func getActivityMetadata() *activity.Metadata {

	if activityMetadata == nil {
		jsonMetadataBytes, err := ioutil.ReadFile("activity.json")
		if err != nil {
			panic("No Json Metadata found for activity.json path")
		}

		activityMetadata = activity.NewMetadata(string(jsonMetadataBytes))
	}

	return activityMetadata
}

func TestCreate(t *testing.T) {

	act := NewActivity(getActivityMetadata())

	if act == nil {
		t.Error("Activity Not Created")
		t.Fail()
		return
	}
}

func TestEval(t *testing.T) {

	defer func() {
		if r := recover(); r != nil {
			t.Failed()
			t.Errorf("panic during execution: %v", r)
		}
	}()

	act := NewActivity(getActivityMetadata())
	tc := test.NewTestActivityContext(getActivityMetadata())

	//setup attrs

	act.Eval(tc)

	//check result attr
}

func TestCalculateDeltas(t *testing.T) {

	defer func() {
		if r := recover(); r != nil {
			t.Failed()
			t.Errorf("panic during execution: %v", r)
		}
	}()

	//setup attrs
	prev := make(map[string]string)
	prev["name"] = "ayman"
	prev["title"] = "E.L.S developer"
	current := make(map[string]string)
	current["name"] = "ayman hassan"
	current["age"] = "32"

	diffs := CalculateDeltas(prev, current)

	//check result attr
	if diffs["title"] != "null" {
		t.Fail()
	}
	if diffs["name"] != "ayman hassan" {
		t.Fail()
	}
	if diffs["age"] != "32" {
		t.Fail()
	}
}
